% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild.R
\name{compute_spatial_window}
\alias{compute_spatial_window}
\title{Compute spatial window for location}
\usage{
compute_spatial_window(locations)
}
\arguments{
\item{locations}{Data frame with lon, lat, radiusx, radiusy}
}
\value{
Data frame with added spatial fields
}
\description{
Converts lon/lat + buffer to:
\itemize{
\item UTM CRS (appropriate zone)
\item Projected extent (xmin, xmax, ymin, ymax)
\item Lonlat extent (lonmin, lonmax, latmin, latmax)
}
}
