% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild.R
\name{define_test_location}
\alias{define_test_location}
\title{Define test location (Noville Peninsula)}
\usage{
define_test_location()
}
\value{
Data frame with single location
}
\description{
Define test location (Noville Peninsula)
}
