% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{fetch_all_pages}
\alias{fetch_all_pages}
\title{Follow rel="next" links, collecting all items}
\usage{
fetch_all_pages(url, max_pages = 100)
}
\arguments{
\item{url}{First-page search URL (from sds::stacit())}

\item{max_pages}{Safety valve}
}
\value{
list(items = list of item lists, n_pages = int)
}
\description{
Follow rel="next" links, collecting all items
}
