% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starc.R
\name{get_assets_from_urls}
\alias{get_assets_from_urls}
\title{Get assets from pre-built query URLs}
\usage{
get_assets_from_urls(query_urls)
}
\arguments{
\item{query_urls}{Query URL(s) from sds::stacit()}
}
\value{
Tibble with asset URLs and metadata
}
\description{
This was born as hrefs from {scene}
}
