% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starc.R
\name{get_assets_single}
\alias{get_assets_single}
\title{Internal: Fetch single STAC query with pagination}
\usage{
get_assets_single(query_url)
}
\arguments{
\item{query_url}{STAC query URL}
}
\value{
Tibble with assets + metadata
}
\description{
This was born as hrefs0 from {scene}
}
