% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild.R
\name{mk_utm_centre}
\alias{mk_utm_centre}
\title{Determine UTM point from lon/lat}
\usage{
mk_utm_centre(lon, lat)
}
\arguments{
\item{lon}{Longitude}

\item{lat}{Latitude}
}
\value{
dataframe with columns x, y in UTM
}
\description{
Determine UTM point from lon/lat
}
