% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild.R
\name{mk_utm_crs}
\alias{mk_utm_crs}
\title{Determine UTM CRS from lon/lat}
\usage{
mk_utm_crs(lon, lat)
}
\arguments{
\item{lon}{Longitude}

\item{lat}{Latitude}
}
\value{
EPSG code as string
}
\description{
Determine UTM CRS from lon/lat
}
