% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest-mgrs.R
\name{month_window}
\alias{month_window}
\title{Month window in RFC3339, inclusive of the month's last second}
\usage{
month_window(month)
}
\description{
Month window in RFC3339, inclusive of the month's last second
}
\keyword{internal}
