% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild.R, R/starc.R
\name{prepare_query}
\alias{prepare_query}
\title{Prepare STAC query}
\usage{
prepare_query(spatial_window, start_date, end_date, collections, provider)

prepare_query(spatial_window, start_date, end_date, collections, provider)
}
\arguments{
\item{spatial_window}{Spatial window with bbox}

\item{start_date}{Start date}

\item{end_date}{End date}
}
\value{
Query specification

Query specification
}
\description{
Prepare STAC query

Prepare STAC query
}
