% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starc.R
\name{process_feature}
\alias{process_feature}
\title{Internal: Extract assets and metadata from single feature}
\usage{
process_feature(feature)
}
\arguments{
\item{feature}{Single STAC feature (one row from features data frame)}
}
\value{
Named list with assets + metadata
}
\description{
Internal: Extract assets and metadata from single feature
}
