% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aatgrid-mgrs.R
\name{utm_extent_to_lonlat}
\alias{utm_extent_to_lonlat}
\title{Lonlat bbox of a UTM extent, densified (helper for cross-zone)}
\usage{
utm_extent_to_lonlat(extent, epsg, n = 21)
}
\description{
Lonlat bbox of a UTM extent, densified (helper for cross-zone)
}
\keyword{internal}
