% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild.R
\name{write_consolidated_assets}
\alias{write_consolidated_assets}
\title{Write consolidated assets to single Parquet file}
\usage{
write_consolidated_assets(
  assets_consolidated,
  output_path = "_targets/assets_consolidated.parquet"
)
}
\arguments{
\item{assets_consolidated}{Consolidated assets tibble}

\item{output_path}{Path to write final Parquet file}
}
\value{
Character. Path to written file
}
\description{
Write consolidated assets to single Parquet file
}
