% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{marker}
\alias{marker}
\title{Marker: max solarday per collection (deduped, derived, never stored)}
\usage{
marker(store = "store")
}
\description{
Marker: max solarday per collection (deduped, derived, never stored)
}
